fetchRealTradeInfo/requirements.txt
aiohttp==3.3.0
cchardet==2.1.1
aiodns==1.1.1
sqlalchemy==1.2.8
xmltodict==0.11.0
PyMySQL==0.8.1
python-dateutil==2.7.3
